# packages/api-types/openapi.yaml
openapi: 3.1.0
info:
  title: TradingAgents API
  version: 1.0.0
  description: |
    Public API gateway for TradingAgents analysis sessions.
    Proxies run execution to the Python agents-service.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /billing/account:
    get:
      tags: [billing]
      summary: Get subscription and usage
      description: |
        Returns the signed-in user's plan and hosted allowance usage breakdown.
        Usage may be sample data while metering is scaffolded.
      operationId: getBillingAccount
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
      responses:
        "200":
          description: Billing account snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingAccountResponse"
        "401":
          description: Missing or invalid auth
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /billing/checkout:
    post:
      tags: [billing]
      summary: Start Managed Payments Checkout
      description: |
        Validates plan selection and creates a Stripe Checkout Session with
        Managed Payments enabled when Stripe is configured. Otherwise returns
        a not_configured payload (and may activate a scaffold subscription when
        authenticated for local UI review).
      operationId: createCheckoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Stripe Checkout Session created; redirect to checkoutUrl
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "501":
          description: Stripe is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "401":
          description: Authentication required when Stripe is configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid checkout payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /billing/subscription/cancel:
    post:
      tags: [billing]
      summary: Cancel subscription at period end
      description: |
        Schedules cancellation at the end of the current billing period
        (`cancel_at_period_end`). Existing runs and shared links are kept.
        New analyses remain available until `accessEndsAt` when status is still
        `active`. Past-due subscribers can also schedule cancel to stop renewals.
      operationId: cancelSubscription
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
      responses:
        "200":
          description: Cancellation scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelSubscriptionResponse"
        "400":
          description: No cancellable subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid auth
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Stripe API failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /config/resolve:
    post:
      tags: [config]
      summary: Resolve config filtered by user credentials
      description: |
        Returns analysts, languages, and only the LLM providers for which
        the user has stored valid credentials. Secret keys are loaded
        server-side and never returned.
      operationId: resolveConfig
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveConfigRequest"
      responses:
        "200":
          description: Filtered configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedConfigResponse"
  /config/providers/{provider}/models:
    post:
      tags: [config]
      summary: List models for a provider (credential-aware)
      operationId: getProviderModelsWithCredentials
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderModelsRequest"
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderModelsResponse"
        "403":
          description: No credentials for provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [config]
      summary: List models for a provider
      operationId: getProviderModels
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: mode
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ModelMode"
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderModelsResponse"
        "404":
          description: Unknown provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      deprecated: true
      schema:
        type: string
      description: Deprecated. Use Authorization Bearer token with a Clerk session JWT instead.
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    BillingPlanId:
      type: string
      enum: [standard, pro]
    BillingInterval:
      type: string
      enum: [monthly, annual]
...
    CheckoutRequest:
      type: object
      required: [planId, interval]
      properties:
        planId:
          $ref: "#/components/schemas/BillingPlanId"
        interval:
          $ref: "#/components/schemas/BillingInterval"
      description: |
        Redirect URLs are intentionally omitted until an origin allowlist is
        implemented for the payment provider integration.
    CheckoutResponse:
      type: object
      required: [status, planId, interval, checkoutUrl, message]
      properties:
        status:
          type: string
          enum: [ready, not_configured]
        planId:
          $ref: "#/components/schemas/BillingPlanId"
        interval:
          $ref: "#/components/schemas/BillingInterval"
        checkoutUrl:
          type: string
          nullable: true
          description: Stripe Checkout URL when status is ready; null when not configured.
        message:
          type: string
        subscriptionActivated:
          type: boolean
    BillingAccountResponse:
      type: object
      required: [subscription, usage, hostedProviderIds, features, agentsModelDisplayName]
      properties:
        subscription:
          type: object
          additionalProperties: true
        usage:
          type: object
          nullable: true
          additionalProperties: true
        hostedProviderIds:
          type: array
          items:
            type: string
        features:
          type: object
          additionalProperties: true
        agentsModelDisplayName:
          type: string
    CancelSubscriptionResponse:
      type: object
      required: [subscription, accessEndsAt]
      properties:
        subscription:
          type: object
          additionalProperties: true
        accessEndsAt:
          type: string
          nullable: true
          format: date-time
          description: When new-run access ends (current period end).
...
    ResearchDepth:
      type: integer
      enum: [1, 3, 5]
    ModelMode:
      type: string
      enum: [all, quick, deep]
...
    ConfigOption:
      type: object
      required: [value, label]
      properties:
        value:
          type: string
        label:
          type: string
    ModelOption:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
        label:
          type: string
        creditMultiplier:
          type: number
          description: Hosted compute-credit multiplier (tokens × this ≈ credits).
        capabilities:
          type: object
          additionalProperties:
            type: boolean
          description: |
            Optional provider-specific capability flags (e.g. anthropicEffort,
            openaiReasoningEffort, googleThinkingLevel).
    ConfigOptions:
      type: object
      required: [analysts, researchDepths, languages, providers]
      properties:
        analysts:
          type: array
          items:
            $ref: "#/components/schemas/ConfigOption"
        researchDepths:
          type: array
          items:
            type: object
            required: [value, label]
            properties:
              value:
                $ref: "#/components/schemas/ResearchDepth"
              label:
                type: string
        languages:
          type: array
          items:
            $ref: "#/components/schemas/ConfigOption"
        providers:
          type: array
          items:
            type: object
            required: [id, label]
            properties:
              id:
                type: string
              label:
                type: string
              backendUrl:
                type: string
                nullable: true
    ProviderModelsResponse:
      type: object
      required: [provider, mode, models]
      properties:
        provider:
          type: string
        mode:
          $ref: "#/components/schemas/ModelMode"
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelOption"
...
    ResolveConfigRequest:
      type: object
      description: Credentials are loaded from server storage; body is ignored.
    ProviderModelsRequest:
      type: object
      required: [mode]
      properties:
        mode:
          $ref: "#/components/schemas/ModelMode"
        providerCredentials:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: string
    CredentialField:
      type: object
      required: [name, label, secret, required]
      properties:
        name:
          type: string
        label:
          type: string
        secret:
          type: boolean
        required:
          type: boolean
        placeholder:
          type: string
          nullable: true
    ProviderCredentialDefinition:
      type: object
      required: [id, label, requiresApiKey, credentialFields, modelSource, apiKeyUrl]
      properties:
        id:
          type: string
        label:
          type: string
        backendUrl:
          type: string
          nullable: true
        requiresApiKey:
          type: boolean
        credentialFields:
          type: array
          items:
            $ref: "#/components/schemas/CredentialField"
        modelSource:
          type: string
          enum: [static, live, static_or_live]
        apiKeyUrl:
          type: string
          nullable: true
          description: Official URL where users can sign up and create an API key.
    CredentialsSchemaResponse:
      type: object
      required: [providers, modelCatalogNote]
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderCredentialDefinition"
        modelCatalogNote:
          type: string
    ResolvedConfigResponse:
      allOf:
        - $ref: "#/components/schemas/ConfigOptions"
        - type: object
          required: [credentialsSchema, availableProviderIds]
          properties:
            credentialsSchema:
              $ref: "#/components/schemas/CredentialsSchemaResponse"
            availableProviderIds:
              type: array
              items:
                type: string
...